Update demo-investigate-gaseous-initialization.cpp
The change in problem specification solves the convergence issue. This is indeed caused by round-off errors. So, adding 1e-6 moles of O2 corrects the bad numerical behavior. The new minimization algorithm implemented in Optima should be able to solve this problem without this minor addition of O2. I still have not had the time to couple new Optima library in Reaktoro. :disappointed: